added scss and moved headernavbar to own folder
