+++ client/src/Header/Header.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import HeaderNavbar from './HeaderNavbar';
+import HeaderNavbar from './Navbar/HeaderNavbar';
 
 class Header extends Component {
   render() {
