+++ client/src/index.css
@@ -1,14 +1,11 @@
 html {
-  height: 100%;
-}
+  height: 100%; }
 
 body {
   margin: 0;
   padding: 0;
   font-family: sans-serif;
-  height: 100%;
-}
+  height: 100%; }
 
 #root {
-  height: 100%;
- }
+  height: 100%; }
